feat: [2u-336] add component, refactor header
