+++ src/course-outline/CourseOutline.jsx
@@ -6,9 +6,9 @@ import { Container, Layout } from '@edx/paragon';
 import SubHeader from '../generic/sub-header/SubHeader';
 import HeaderNavigations from './header-navigations/HeaderNavigations';
 import OutlineSideBar from './outline-sidebar/OutlineSidebar';
-import messages from './messages';
 import { useCourseOutline } from './hooks';
-import OutlineSection from './outline-section/OutlineSection';
+import SectionCard from './section-card/SectionCard';
+import messages from './messages';
 
 const CourseOutline = ({ courseId }) => {
   const intl = useIntl();
@@ -54,7 +54,7 @@ const CourseOutline = ({ courseId }) => {
               <div>
                 <section className="course-outline-section">
                   {sectionsList.length ? sectionsList.map((section) => (
-                    <OutlineSection section={section} />
+                    <SectionCard section={section} />
                   )) : null}
                   {/* TODO add status bar and list of outlines */}
                 </section>
