+++ src/course-outline/CourseOutline.scss
@@ -1,3 +1,4 @@
 @import "./header-navigations/HeaderNavigations";
 @import "./outline-sidebar/OulineSidebar";
-@import "./outline-section/OutlineSection";
+@import "./section-card/SectionCard";
+@import "./card-header/CardHeader";
