--- src/course-outline/outline-section/OutlineSection.scss
@@ -1,66 +1,0 @@
-.outline-section {
-  @include pgn-box-shadow(1, "centered");
-
-  padding: 1.5rem;
-  cursor: move;
-  margin-bottom: 1.5rem;
-
-  .outline-section__header {
-    display: flex;
-    align-items: center;
-
-    .header-expanded-btn {
-      justify-content: flex-start;
-      padding: 0;
-      width: 80%;
-      height: 1.5rem;
-      margin-right: .25rem;
-      background: transparent;
-
-      &::before {
-        display: none;
-      }
-
-      & svg {
-        width: 1.5rem;
-        height: 1.5rem;
-      }
-
-      &.collapsed .pgn__icon {
-        transform: rotate(180deg);
-      }
-    }
-
-    .header-badge-status {
-      display: flex;
-      padding: 1px .5rem;
-      justify-content: center;
-      align-items: center;
-      gap: .25rem;
-      border-radius: .375rem;
-      border: 1px solid $light-300;
-      margin: 0 .75rem;
-    }
-
-    .header-menu {
-      display: flex;
-      align-items: center;
-      width: 1.5rem;
-      height: 1.5rem;
-    }
-  }
-
-  .outline-section__content {
-    margin-top: 1.75rem;
-  }
-}
-
-.header-tooltip {
-  .tooltip-inner {
-    max-width: 18.75rem;
-  }
-
-  .arrow {
-    transform: translate(5.75rem, 0) !important;
-  }
-}
