+++ src/course-outline/section-card/SectionCard.scss
@@ -1,0 +1,11 @@
+.section-card {
+  @include pgn-box-shadow(1, "centered");
+
+  padding: 1.5rem;
+  cursor: move;
+  margin-bottom: 1.5rem;
+
+  .section-card__content {
+    margin-top: 1.75rem;
+  }
+}
