+++ src/i18n/messages/it.json
@@ -748,5 +748,16 @@
   "course-authoring.course-outline.sidebar.section-4.descriptions-2.hide": "Hide",
   "course-authoring.course-outline.sidebar.section-4.descriptions-3": "To hide the content of a subsection from learners after the subsection due date has passed, select the Configure icon for a subsection, then select {hide}. Grades for the subsection remain included in grade calculations.",
   "course-authoring.course-outline.sidebar.section-4.descriptions-3.hide": "Hide content after due date",
-  "course-authoring.course-outline.sidebar.section-4.link": "Learn more about content visibility settings"
+  "course-authoring.course-outline.sidebar.section-4.link": "Learn more about content visibility settings",
+  "course-authoring.course-outline.section.expandTooltip": "Collapse/Expand this section",
+  "course-authoring.course-outline.section.status-badge.live": "Live",
+  "course-authoring.course-outline.section.status-badge.published-not-live": "Published not live",
+  "course-authoring.course-outline.section.status-badge.staff-only": "Staff only",
+  "course-authoring.course-outline.section.status-badge.draft": "Draft",
+  "course-authoring.course-outline.section.menu.edit": "Edit",
+  "course-authoring.course-outline.section.menu.publish": "Publish",
+  "course-authoring.course-outline.section.menu.configure": "Configure",
+  "course-authoring.course-outline.section.menu.duplicate": "Duplicate",
+  "course-authoring.course-outline.section.menu.delete": "Delete",
+  "course-authoring.course-outline.section.menu.new-subsection": "New Subsection"
 }
